Revert "80 mobiledesktop loading update size"
