+++ src/components/loading/Loading.tsx
@@ -3,8 +3,57 @@ import anime from 'animejs';
 import './loading.scss';
 import '../../scss/_variables.scss';
 import arrow from '../../assets/icons/arrowDown.svg';
-import { Vector } from './vector';
 
+class Vector {
+    x: number;
+    y: number;
+    constructor({ x = 0, y = 0 }) {
+        this.x = x;
+        this.y = y;
+    }
+    add(...vectors: Vector[]) {
+        return vectors.reduce(
+            (acc, curr) => ( new Vector({
+                x: acc.x + curr.x,
+                y: acc.y + curr.y,
+            })),
+            new Vector({
+                x: this.x,
+                y: this.y,
+            })
+        );
+    }
+  
+    sub(p2: { x: number; y: number; }) {
+        return new Vector({
+            x: this.x - p2.x,
+            y: this.y - p2.y,
+        });
+    }
+  
+    mul(n: number) {
+        return new Vector ({
+            x: this.x*n,
+            y: this.y*n,
+        });
+    }
+  
+    dot(v2: this) {
+        return this.x*v2.x + this.y*v2.y;
+    }
+  
+    len() {
+        return Math.sqrt(this.dot(this));
+    }
+    
+    norm() {
+      let len = this.len();
+      return new Vector({
+        x: this.x/len,
+        y: this.y/len,
+      });
+    }
+}
 
 function Loading(this: any) {
     // === Defining data ===
@@ -26,7 +75,7 @@ function Loading(this: any) {
             setCanvasWidth(wrapperWidth);
             setCanvasHeight(wrapperHeight);
         }
-    });
+    }, );
     
     // === Defines canvas ===
     // Note: not using useRef(); because I haven't figured out how to get it as the correct type; HTMLCanvasElement.
@@ -144,8 +193,6 @@ function Loading(this: any) {
         points = step(points, originalPoints);
         draw(points, ctx);
         t++;
-        setCanvasWidth(wrapperWidth);
-        setCanvasHeight(wrapperHeight);
         window.requestAnimationFrame(animateIt);
     }
     window.requestAnimationFrame(animateIt);
